BREAKING CHANGE: movendo página '/realms' para '/'
